backend/pyproject.toml
[project]
name = "historical-travel-agent-backend"
version = "0.1.0"
description = "歴史学習特化型旅行AIエージェント - バックエンド"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy>=2.0.36",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.10",
    "redis>=5.2.0",
    "google-cloud-aiplatform>=1.70.0",
    "google-cloud-storage>=2.18.0",
    "python-multipart>=0.0.20",
    "pillow>=11.0.0",
    "aiofiles>=25.1.0",
    "filetype>=1.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "hypothesis>=6.122.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "pyright>=1.1.300",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "hypothesis>=6.122.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "pyright>=1.1.300",
]
